requirements.txt
google-generativeai==0.5.4
pillow==10.3.0
streamlit==1.33.0